requirements.txt
# Core AI/LLM Libraries
anthropic>=0.18.0
openai>=1.12.0
langchain>=0.1.0
langchain-community>=0.0.20
tiktoken>=0.5.2

# Web Search & Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
duckduckgo-search>=4.4.0
newspaper3k>=0.2.8
selenium>=4.16.0
playwright>=1.41.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Storage & Vector Database (Optional)
chromadb>=0.4.22
faiss-cpu>=1.7.4

# CLI Interface
click>=8.1.7
rich>=13.7.0
colorama>=0.4.6

# Utilities
python-dateutil>=2.8.2
aiohttp>=3.9.0
tenacity>=8.2.3

# Development Tools (Optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=24.1.0
pylint>=3.0.0
flake8>=7.0.0
